+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Next
+- [feat] [Crontab friendly generator](https://it-tools.tech/crontab-generator)
+
 ## 1.6.0
 - [feat] [BIP39 generator](https://it-tools.tech/bip39-generator)
 - [feat] [Base 64 converter](https://it-tools.tech/base64-string-converter)